feat(api): add server metrics endpoint and simplify service history response
- Add /server/:id/metrics API for querying TSDB server metrics
- Simplify getServiceHistory by removing redundant data conversion
- Change AvgDelay type from float32 to float64
- Remove generated swagger docs (to be regenerated)
- Update TSDB query, writer and tests

+++ model/service_api.go
@@ -26,7 +26,7 @@ type ServiceResponseItem struct {
 	CurrentDown uint64       `json:"current_down"`
 	TotalUp     uint64       `json:"total_up"`
 	TotalDown   uint64       `json:"total_down"`
-	Delay       *[30]float32 `json:"delay,omitempty"`
+	Delay       *[30]float64 `json:"delay,omitempty"`
 	Up          *[30]uint64  `json:"up,omitempty"`
 	Down        *[30]uint64  `json:"down,omitempty"`
 }
